codigo/requirements.txt
SpeechRecognition
pyaudio
pyttsx3
sounddevice
gtts
pygame
#threading
#os
#sys
pytz
requests
#random
openai
geopy
timezonefinder
googletrans==4.0.0-rc1
secure-smtplib
tuya-connector-python
tuya-iot-py-sdk
psycopg2



